PR comments, moved complete and template strings out into it's own class
